refactor: define authSpace functions in store instead of separating them
